// src/types/profile.ts
export interface UserProfileResponse {
	user_id: number;
	name: string;
	level: number;
	gender: string;
	age: number;
	rank: string;
	property: string;
	signup: string;
	awards: number;
	friends: number;
	enemies: number;
	forum_posts: number;
	karma: number;
	role: string;
	donator: number;
	player_id: number;
	status: string;
	last_action: LastAction;
	faction: Faction;
	job: Job;
	married: Married;
	profile: string;
}

export interface LastAction {
	status: string;
	timestamp: number;
}

export interface Faction {
	faction_id: number;
	faction_name: string;
	position: string;
	days_in_faction: number;
}

export interface Job {
	company_id: number;
	company_name: string;
	position: string;
	days_in_company: number;
}

export interface Married {
	spouse_id: number;
	spouse_name: string;
	duration: number;
}